Move split_receipt module into its own file
